[feat] switch to newer plantuml
We switched to a newer plantuml. It seems that most plantuml setups
use .jar files instead of executables (the Ubuntu PPA exec is fairly
outdated).

In doing so, we placed these files in ~/Code/others/plantuml/, and
updated the config.el file to use this.

Bonus: we set org to hide code blocks by default when opening a file.