[feat] add venv support to emacs (part 1)
Previously, I had been using poetry as for virtual environments.
This is very well supported by emacs (and doom-emacs). The downside
is that poetry does not support using or importing projects that
have been set up with a different builder/config (e.g. setup.py,
requirements.txt, non-poetry pyproject.toml).

A (potentially good) alternative to this is PDM. It follows many
of the same concepts/ideals of poetry, but also appears to have
a proper import tool. The downside is just that emacs/doom-emacs
does not natively support it.

This would *replace* poetry. Note that this is only a bare-bones
replacement. To truly get it up-to-snuff, we would also need
LSP/DAP working with venv. See:
https://old.reddit.com/r/emacs/comments/n1gkkk/doom_emacs_pyright_lsp_conda/