[feat] setup dap-mode properly per-project
For some reason, dap-mode was not working by default on my work pc.
The projectile project root was not being detected/handled by default.
I am working around this by registering dap-debug templates with
(projectile-project-root) as the cwd.

Unfortunately, these must be evaluated *at the time*
dap-register-debug-template is called. Because of this, we wrap it in
a function and call it on workspace switching.