feat: try to avoid poetry infinite-loop
There is a possible 'infinite' loop that can happen when opening a
project that is managed by poetry. It causes poetry to seem to be
reloaded forever (I believe at some point it stops, but it takes a
looong time). The changes here try to properly link poetry to the
appropriate mode hooks.